feat(httpsdk): Adjust code structure
