+++ .gitignore
@@ -1,6 +1,5 @@
 tools/
 .idea/
-protoc-gen-go-httpsdk
 .fleet/
 
 dist/
