+++ .goreleaser.yaml
@@ -13,6 +13,7 @@ builds:
       - linux
       - windows
       - darwin
+    main: ./cmd/protoc-gen-go-httpsdk
 archives:
   - replacements:
       darwin: Darwin
