+++ internal/gateway/classify_scope.go
@@ -1,12 +1,12 @@
-package internal
+package gateway
 
 import (
 	"github.com/jaronnie/protoc-gen-go-httpsdk/internal/vars"
 	"github.com/jaronnie/protoc-gen-go-httpsdk/utilx"
 )
 
-// getAllScopeVersionsMap {"corev1":"CoreV1", "corev2":"CoreV2"}
-func getAllScopeVersionsMap(scopeResourceGws vars.ScopeResourceGateway) map[string]string {
+// GetAllScopeVersionsMap {"corev1":"CoreV1", "corev2":"CoreV2"}
+func GetAllScopeVersionsMap(scopeResourceGws vars.ScopeResourceGateway) map[string]string {
 	scopeVersionMap := make(map[string]string, 0)
 
 	for k, _ := range scopeResourceGws {
